ai_node/requirements.txt
# Core LangChain
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.31    # <- change this line
langchain-text-splitters==0.3.5
langchain-ollama==0.2.2

# Vector Store
faiss-cpu==1.9.0

# PDF Processing
pypdf==4.1.0
pdfplumber==0.11.0
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.4.0

# Ollama
ollama>=0.4.4,<1.0.0

# Text Processing
nltk==3.9.1

# Utilities
python-dotenv==1.0.1

# Optional: for GPU support
# torch==2.1.0
# torchvision==0.16.0

# API Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12